ocr-service/requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
Pillow
python-doctr[torch]
pix2text
